+++ modules/login/lib/components/user/user.module.scss
@@ -1,0 +1,18 @@
+.username {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  overflow: hidden;
+
+  img {
+    width: 150px;
+    height: 150px;
+    border-radius: 100%;
+  }
+
+  a {
+    margin-top: -15px;
+    transform: none !important;
+  }
+}
